Created service classes and interface

+++ src/main/java/za/ac/cput/Repository/VehicleRepository.java
@@ -3,5 +3,5 @@ package za.ac.cput.Repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import za.ac.cput.Domain.registration.Vehicle;
 
-public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
+public interface VehicleRepository extends JpaRepository<Vehicle, Integer> {
 }
